finish task 0. The basics of async

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -8,7 +8,7 @@ import random
 import asyncio
 
 
-async def wait_random(max_delay: int = 10) -> int:
+async def wait_random(max_delay: int = 10) -> float:
     """
     coroutine that takes an integer max delay and waits for
     delay between 0-max_delay and eventually returns the delay time
